Break setNotebook into action & epic
